requirements.txt
django==3.2
django-ninja==0.19.1
psycopg2==2.7.7
django-ninja-jwt
django-redis
django-Ninja-Extra